use resolved hostname for all metrics (#11634)
When DBM is enabled we want all telemetry, both standard metrics as well as DBM events, to have the hostname set to the resolved hostname. This ensures that when an agent is monitoring a remote host (i.e. RDS or Azure managed) all SQL Server telemetry is tagged with the hostname of the remote host and not the agent's hostname.