Clarify the instructions releasing integration
Step 2 was a bit unclear. The instructions made it seem like the integration release pipeline triggering can be done in parallel with the creation of the PR. I moved it to its own step to hopefully make it clearer that the release pipeline should be triggered after the PR has been merged in.