sqlserver fix missing object_name for index fragmentation metrics
When `database_autodiscovery: True` then `object_name` was always `none` for any objects that came from a database other than the default database.

This is because [object_name](https://docs.microsoft.com/en-us/sql/t-sql/functions/object-name-transact-sql?view=sql-server-ver15) defaults to assuming it's in the current database context if not specified.

Now we update the query to always specify database_id to ensure we can correctly resolve the object name.