Add note about special characters in property name (#11698)
* Add note about special characters in property name

Some users will find that the service name includes a special character - MSSQL$CRMAWS - in this instance when they enter this into the configuration file just as it is, the windows service check will return unknown. What's happening is that the $ gets turned into an underscore and the agent can't find it since that isn't its name. They need to escape special characters like this - MSSQL\$CRMAWS for the service name to be properly found. Since this isn't mentioned anywhere in this doc, there only option is to ask support why its messed up.

* Update windows_service/README.md

* Update windows_service/README.md

* Move README link to bottom

Co-authored-by: cswatt <cecilia.watt@datadoghq.com>
Co-authored-by: Andrew Zhang <andrew.zhang@datadoghq.com>