remove extra : in the value field for `<INSTANCE_CONFIG>` (#11852)
I worked with a customer recently who copied the value in the instance config section. He then followed the auto-discovery integration template [5] to find how he should setup the config file. The extra : that he had copied was causing his configuration not to function. After removing the : he stopped receiving errors. I'm not sure if removing the : is the best approach to avoid any confusion like this but wanted to bring it your attention just to get the ball rolling.