remove execution plan user attribute
Remove the `user_name` attribute from query metrics and execution plans.

This user attribute has turned out to be confusing to use because it doesn't represent the user that actually executed the query. Instead, it represents the following (from [the docs](https://docs.microsoft.com/en-us/sql/relational-databases/system-dynamic-management-views/sys-dm-exec-plan-attributes-transact-sql)): *Value of -2 indicates that the batch submitted does not depend on implicit name resolution and can be shared among different users. This is the preferred method. Any other value represents the user ID of the user submitting the query in the database.*. To remove the confusion we are removing this attribute.

By removing the plan user attribute the statement metrics query also becomes less expensive because we no longer need to aggregate by the plan user and we no longer need to join on `sysusers`.