handle encrypted plan cache entries
Include query stats & plan samples for encrypted plan cache entires:

* read out the `encrypted` field from both `sys.dm_exec_sql_text` and `sys.dm_exec_query_plan`
* allow `None` values for `query_signature`, `statement`, `obfuscated_statement`, `db.plan.definition` when we know it's encrypted

This also fixes the `collection_errors` message which used to say `can only parse strings` because it used to try to obfuscate a NULL plan. Now there will be no collection error as encrypted plans are expected to be NULL, and the `is_encrypted` field will make it clear that the plan or statement are missing because it's encrypted.