Change log level to debug
This is a very noisy log.  In our environment, we wish to monitor some specific process which exists in kubernetes masters only, but since there's no way to target this check to a specific nodes, this log spams on thousands of nodes where it is expected that the process doesn't exist. This should be a debug level log.