Implement Multi-instance counters without relying on Windows PdhEnumObject(refresh=TRUE) API.
Removed non-default edge case support of dynamic addition/removal of configured
performance counters. Even if a provider of a performance counter object,
specified in our configuration, is not running, as long as its provider has
been installed/deployed on the box, that is all needed for this integration.
When the counter provider starts, Agent will discover its new instances and
report them as metrics correspondingly.

However, if the provider is installed AFTER the agent starts running, this
implementation will not be able to automatically discover and use it. In
some way it is a step back because this ability had been added in 7.34
among few other very important changes including detection of appearance
and disappearance of performance counter instances, support for non-unique
instances and  localization support. In this release detection of newly
registered performance counters providers (and their objects and counters)
is pulled out for the following reasons:
* It does not appear that the feature was needed by customers
* It adds non-insignificant complexity to the code
* It relies on very slow and resource hungry PdhEnumObjects(refresh=TRUE)
  function.
* It pretty much requires the agent process to be run as administrator or
  local system user (otherwise will generate a handful of error messages
  in the event log every time refresh run). Microsoft support confirmed
  it even though it is not documented. It is difficult to explain to
  customers.
* The logic of dynamically adding and removing counters handles to
  facilitate support appears to may cause memory leaks in some counters
  (documented at least for ASP.NET e.g.).

Perhaps in future if customers would really want that and are ready to
deal with the caveats above we can potentially add it back.