kafka_consumer send partitions as histogram
We have kafka multiple clusters with > 50 partitions, > 100 topics and > 200 consumer groups. That ~~adds up~~ multiplies up in number of metrics. We are not particularly interested in individual partitions so this feature submits them as a histogram to reduce the total number of metrics.

Please let me know your thoughts/suggestions.