reduce overhead of integration tracing
When `integration_tracing: true` we will now trace only each integration's `run` method instead of tracing all methods. All methods are traced only when the new `integration_tracing_exhaustive: true` option is set.

When the `DDEV_TRACE_ENABLED` env var is set it turns on both `integration_tracing` and `integration_tracing_exhaustive`.

We do this in order to reduce the overhead of tracing when running continuously in production. Some integrations make 1000s of calls to various internal methods, leading to 1000s of spans being recorded. Many of these methods are so fast that the overhead added by tracing starts to exceed the execution time of the methods themselves.