improve integration tracing of warnings & errors
When `integration_tracing: true` we now trace all `AgentCheck.warning` calls, ensuring a detailed error span is collected which includes the warning message as well as the full stack trace at the point the warning was recorded.

We also now include `AgentCheck.check` in the set of methods that are traced when `integration_tracing: true`. If `AgentCheck.check` throws an exception then it is handled in `AgentCheck.run` where the tracing library doesn't know about it. By tracing `AgentCheck.check` we ensure any unhandled exceptions are captured in the trace.