Continue directory traversal after errors (#14587)
* Continue traversing directories after errors occur

* Remove dead error-handling path

Now that errors are handled via a callback to `walk`, we no longer
need to handle them from the outer iterator.

* Expand the explanation on why we use a custom walk function

The original docstring did somewhat reflect the issue, but it wasn't
enough for us to remember why we needed this and why we aren't using
the stdlib's function without doing some research. Hopefully this will
help us remember in the future.

* Restore folder permissions on test to allow for cleanup

* Changes to support python 2

Turns out py2's `os.chmod` function doesn't support passing the mode
as a named argument.