Reduce number of consumer creations for highwater offset collection (#15476)
- Remove optimization based on consume group and cluster id
- Explicitly close consumer instance

Context: Optimization based on consumer group and cluster id helped to reduce
number of broker calls, but it still created multiple consumer instances

As the consumer groups are collected based on kafka_connect_str and the
same is used for creating consumer at the time of highwater offset
collection, they both are bound to be part of the same cluster. This
means that highwater offset for (topic, partition) will be the same as
the check runs in the context of a single cluster. Therefore it is
beneficial to optimize based on (topic, partition) instead of
consumer group or cluster id.