Remove directConnection parameter
According to the mongo docs here https://www.mongodb.com/docs/mongodb-shell/connect/#connect-to-a-replica-set this parameter gets added anyway and it seems from this PR https://github.com/DataDog/integrations-core/pull/13454 that we are supposed to be able to specify this option but it doesn't seem like this is actually possible. Removing this from my code on the machine that runs my agent worked well for me - so it definitely is not needed in some setups.