Update README.md
Currently the DBM docs imply that we cannot use autodiscovery, and that ports must be static. this is often not the case in enterprise DBs, as ports are dynamically set and autodiscovered. 
per https://github.com/DataDog/integrations-core/pull/13135 we can support dynamic port autodiscovery and our docs should reflect that more clearly so as not to disuade DBAs