feat(php_fpm): track maximum active processes (#16193)
* feat(php_fpm): track maximum active processes

The PHP-FPM status page also provides the "max active processes" statistic, which translates to the maximum number
of processes that can be executed at once. In the PHP-FPM configuration, this is controlled by the pm.max_children
key. This is particularly useful in understanding saturation within the PHP-FPM container, in a manner that allows
horizontal scaling in a more intelligent way than the built-in CPU and memory metrics in Kubernetes. This pairs
well with the Datadog External Metrics functionality as well

* chore: add changelog.d entry for new metric