Fix vacuum age computation
age(t1) will yield `t1 - current_date at midgnight` which is incorrect for
our usage.
We want the number of seconds since the last vacuum event. We will use
age(t1, t2) which yield `t1 - t2` using the start of the statement time
as t1 and the vacuum event as t2.