PostgreSQL: Optimise table count query
We currently rely on pg_stat_user_tables to count the number of tables
per schema. With a high number of tables, this view is expensive to query
due to the order by and a lot of work is wasted as we ignore all toast
tables and system schema like pg_catalog and information_schema.

We can replace this by a direct call to pg_class, filtering as much as
possible early and keeping the price of pg_namespace join low by pushing
the join to the end.