Postgres Performance Optimization: Cache baseline metrics and apply updates incrementaly (#17554)
* Revert "Revert "[SDBM-842] Postgres integration performance optimization: Limit how many records are pulled from pg_stat_statements. (#17187)" (#17397)"

This reverts commit db9b87e0fc6f7f0bc352f9fe70f10fb07cd9d8fa.

Introduce a new 'baseline_metrics' cache which provides the baseline metric values for metric calculations. It's effectively a cache of pg_stat_statements, minus the query text which can be large and expensive to query / store. Once this baseline cache is in place, subsequent check runs only fetch metrics for the queries that have been called since the last run. This greatly improves the amount of data we have to return to the agent, as well as how many metrics rows we have to process.