Remove pg_stat_user_tables call and add toast relation metrics
pg_stat_user_tables view is inefficient on databases with a lot of
tables, possibly leading to temporary bytes being read and written. To
avoid this, we can directly use the pg_stat_* functions.

We also add metrics from the toast tables like live rows and number of
rows fetched.