Make collection interval configurable on a per query basis
Some queries executed by the agent can be expensive to run, like getting
tables' sizes on db with a lot of relations. This could lead to
situation where running the check takes longer than the collection
interval. Currently, the only solution is to disable the offending which
is not great.

This patch adds a way to make collection_interval configurable for
queries that are using the QueryExecutor.