Add cluster_name as tag to PG metrics (#18402)
* Move load_system_identifier to check_initialization

system_identifier will never change within the lifetime of a cluster so
we can move it in initialization's checks

* Add cluster_name as metric tag if set

PostgreSQL instances can set a cluster_name parameter that will be used
in the process' title. This can be used to identify a cluster that
doesn't rely on custom tags. This patch adds the cluster_name as a
metric tag if it was set in the configuration file.