Report pg_class xmin
Add a postgresql.relation.xmin reporting the xmin of the relation's
row in pg_class. This will allow to detect if and when a relation was
modified through a DDL as most of DDL will modify the relation's
pg_class.

Other processes may modify pg_class, like ANALYZE to update tuple
estimations. However, those updates are done in place and won't bump the
row's xmin.