Fix CouchDB 2+ message_queues metric collection (#19521)
* Fix CouchDB 2+ message_queues metric collection

At the moment, the `count` value in the metric block is recorded
as a gauge. That value is the number of instances of each type
of message queue. Say you have two database files open, then
`message_queues.couch_file.count` will be `2`. That information
is already available in the `_stats` endpoint as `open_os_files`.

The actually valuable part of the message_queue metric is the
`max` entry, it shows that the longest backlog of messages for
each instance type. A longer backlog means a reduction in
performance. Without this patch, this value is not available.

Technically all of the `max`, `min`, `50`, `90`, `99` are
valuable, but I wanted to keep this patch as small as possible
until it is clear the change will be accepted.

I’m happy to add the other fields as well if desired.

* OPTIONAL: expose more useful message_queues metrics

* lint

* fix: add assertions for new stats

* add new message queue metric descriptions

* changelog

* move changelog

* delete

---------

Co-authored-by: steveny91 <steven.yuen@datadoghq.com>