fix: allow EXPLAIN on multi-statement SQL beginning with SET (#20106)
* feat: allow EXPLAIN on multi-statement SQL beginning with SET

SQL strings containing multiple statements cause problems with the ex-
isting explain logic. This change adds a regex to strip an arbitrary
number of `SET` commands from the head of all incoming SQL, such that
the eventual `EXPLAIN` will be performed on just the trailing SQL.

This doesn't fully address all multi-statement SQL strings: in part-
icular clients might send e.g. both a SELECT and an UPDATE in the same
string. As before, this would be passed as-is to the EXPLAIN machinery.

Refs: DBMON-2626

* Add changelog entry for new EXPLAIN behavior

* Simplified regex

* Address code review feedback

* This is more of a fix than feature

* Add integration tests, fix logic

* Refactor

* Formatting fix

* Handle leading comment