Fix set explain regex (#20319)
* Reapply "fix: allow EXPLAIN on multi-statement SQL beginning with SET (#20106)" (#20282)

This reverts commit 6486c7f493ace83c81ead34360634a2a64609d17.

* Address catastrophic backtracking in EXPLAIN regex

Took a bit to pin down why this was happening, but with a fuzzer and
some articles about exponential regex performance, I could address the
underlying issue without changing the code much.

* Fix changelog

* Formatting fixes

* Pass query_signature down into explain_statement

No need to recalculate this, and passing it down will allow trimmed SQL
to reuse the original query's signature.