Preserve source for Traefik status codes
The Traefik log fields `OriginStatus` and `DownstreamStatus` are currently both merged into `http.status_code`, and the source fields are dropped.

However, origin and downstream status do not represent the same data, and merging them into each other like this reduces the fidelity of log data. In particular, it makes it impossible to disambiguate the source of an error - e.g., I cannot tell whether Traefik's target returned a 503, or whether Traefik failed to connect to the target in the first place.

I should note that this is currently causing production pain for us by making it difficult to create useful metrics tracking Traefik's health; and there is no available metric in Traefik which otherwise disambiguates this data.

This change would keep the remapping for compatibility, but would preserve the source fields to ensure sufficient fidelity is preserved.