PG: Directly run ANALYZE for the test_analyze_progress
Tests are currently using `vacuum analyze` to check metrics reporting
analyze progress. If the vacuum phase takes more than 1s, the test will
fail as the wait_for_value will exit and only a running vacuum will be
present when metrics are collected.
To avoid flakiness, we skip the vacuum phase and directly run an analyze.