feat: Fix Ceph version lookup for newer releases
Ceph made various breaking changes to the status output in the `octopus`
release in 2020, including (but not limited to) the entire `mon_status`
command being removed.

The DataDog check as-was only checked if the release was exactly
'octopus', not any later release as well. Given that `octopus` is now 5
years old, and there doesn't seem to be anything in the responses that
gives a semantic version of similar to numerically compare, it's easiest
to just _assume_ that we will get stats the New Way :tm:, and also try
any old way if the `mon_status` content exists in the `raw` map and the
new way failed.