PG: Handle locks without relation like transactionid locks
In postgres, a transaction can wait for another transaction to end by trying
to acquire a sharelock on the target transactionid's lock. We currently
exclude those due to the join on relname/schema/db as they have null
values.

This PR convert pg_locks to a dynamic query and handles locks with null
relname/schema/db.