Update README.md (#21299)
* Update README.md

Update password note as per feedback from customer.
I'd like to inform you that we were finally able to setup the Redis integration using the annotation with Redis support. The structure of the annotation, the endpoint and port were correct but the password apparently was supposed to be the password for the 'default' user. We use in our application (the Pod running in EKS where we place the annotation) different credentials, specific user with access to the DB, but apparently for this integration, the 'default' is needed.

* Update README.md

* Update README.md

---------

Co-authored-by: Ida Adjivon <65119712+iadjivon@users.noreply.github.com>