Ensure that ddev does not treat worktrees as extra packages (#20444)
### What does this PR do?
Updates ddev to ensure that when scaning packages in the repo, any git worktree found is not treated as a possible integration.

### Motivation
When working with worktrees it is common to define the worktree as a folder within your repository directory. When ddev scans pacakges within the repo it checks all directory as possible integrations/packages, this includes worktrees folders. This triggers undesired side effect like slow linting and format checks. This is because `ddev test -s` also runs on untracked items and worktrees will likely be untracked.

### Review checklist (to be filled by reviewers)

- [ ] Feature or bugfix MUST have appropriate tests (unit, integration, e2e)
- [x] Add the `qa/skip-qa` label if the PR doesn't need to be tested during QA.
- [ ] If you need to backport this PR to another branch, you can add the `backport/<branch-name>` label to the PR and it will automatically open a backport PR once this one is merged