PG: Add metrics for pg_stat_replication's sent/write/flush/replay (#21844)
* PG: Add metrics for pg_stat_replication's sent/write/flush/replay

pg_stat_replication provides metrics on the last sent/write/flush/replay
WAL location by a standby server.

sent delay doesn't depend on a feedback message from the standby since
it tracks the sent WAL through the connection. This can be used to
gauge how fast and how late a standby is when catching up

* Make PG checkpoint test less flaky

Only run checkpoint and check metrics on the primary. This will remove
the possible uncertainty of having the Checkpoint record being correctly
propagated to the standby before standby's checkpoint is triggered.

* Reduce test_backend_transaction_age flakiness

Add the timestamp to the application name to reduce the risk of leftover
query and connection being present when metrics are collected.