Update postgresql.yaml
Removing a log sample from the first Grok Parser of the Postgresql Integration Pipeline. 

Grok Parsers allow a maximum of 5 log samples but this Grok Parser currently contains 6 samples. If a user clones this pipeline and edits any processor within it, they will be unable to save changes to the pipeline.

The sample that is removed was listed as the second sample but falls under the same rule as the first sample (default_format)