Correct row count calculation. (#22055)
* Correct row count calculation.

Currently because we're joining with indexes, one table produces one row per index before grouping. Summing the row counts across all of these rows leads to an inflated row count (the actual row count x the number of indexes). Taking the MAX ensures we just get the singular row count, and it's ok to do since all of the row counts in the group will be the same.

* Simplify test by only asserting row count.

* Add changelog.

* correct pr number.

* Lint fixes.